+++ packages/miniapp-builder-shared/src/index.js
@@ -1,19 +1,22 @@
-const filterNativePages = require('./filterNativePages');
-const getAppConfig = require('./getAppConfig');
+const separateNativeRoutes = require('./separateNativeRoutes');
+const normalizeStaticConfig = require('./normalizeStaticConfig');
 const getPluginConfig = require('./getPluginConfig');
 const pathHelper = require('./pathHelper');
 const platformMap = require('./platformMap');
 const constants = require('./constants');
 const autoInstallNpm = require('./autoInstallNpm');
 const checkIsMiniappPlatform = require('./checkIsMiniappPlatform');
+const { transformAppConfig, transformPageConfig } = require('./transformConfig');
 
 module.exports = {
-  filterNativePages,
-  getAppConfig,
+  separateNativeRoutes,
+  normalizeStaticConfig,
   getPluginConfig,
   pathHelper,
   platformMap,
   constants,
   autoInstallNpm,
-  checkIsMiniappPlatform
+  checkIsMiniappPlatform,
+  transformAppConfig,
+  transformPageConfig
 };