+++ packages/miniapp-webview-config/src/types.ts
@@ -1,0 +1,10 @@
+export interface AppItemType {
+  entryName: string;
+  url?: string;
+  source: string;
+  name?: string
+}
+
+export interface AppConfigType {
+  routes: AppItemType[]
+}