+++ packages/jsx2mp-loader/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.4.33] - 2021-12-14
+
+### Changed
+
+- update miniapp-builder-shared to v3.0.0
+
+## [0.4.32] - 2021-11-03
+
+### Fixed
+
+- Can't find app_css custom component when the file and app_css are in the same directory
+
+## [0.4.31] - 2021-11-01
+
+### Changed
+
+- Use @babel/preset-env for js/ts files in dev mode
+
 ## [0.4.30] - 2021-08-31
 
 ### Fixed