+++ packages/miniapp-builder-shared/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.3.0] - 2021-12-14
+
+### Changed
+
+- `normalizeStaticConfig` instead of `getAppConfig`
+- `separateNativeRoutes` instead of `filterNativePages`
+
+## [0.2.12] - 2021-11-02
+
+### Fixed
+
+- Use CommonJS to export checkAliInternal method
+
+## [0.2.11] - 2021-11-01
+
+### Changed
+
+- Use self-built method to check ali internal network
+
 ## [0.2.10] - 2021-08-05
 
 ### Changed